First draft of API
Includes 2 APIs: http, which is an 1-1 replication of the http api; and the user one, which adds some syntactic sugar on top of the http and makes things easier to use.

Test suite gives an idea of how things work.

This is work in progress, so there is no documentation.