Fix Tests for Importing non-csv Files
* Change all instances of "seperator" to "separator"

There were multiple instances of this typo both in code and in comments
which have now been fixed.

* Fix tsv tests

All of the tests for tsv files were pointing to the .csv file. In
addition, that tsv file's content was a csv. The tests have been changed
to use the tsv and the tsv is now a tab separated values file. An
additional back slash was added to _IMPORT_EXTENSIONS to properly
escape "\t" and not have it be parsed as a tab.

* Add test for custom delimiters

I added a test for a custom delimiter, in this case the pipe character
"|" because that was not working as expected before.

* Fix overwriting the options dict

In admin.py there was a branch of code which is supposed append
the options dict but instead would overwrite it. This will now produce
the expected behavior.

* Fix inconsistent argument placement in docs

The constructor for ImportFile had separator as the 5th positional
argument in the code but 3rd in the docs. Now it is 5th in the docs
to match the code

* Import pytest in test_cache.py

This was added to make debugging tests in vscode work