Update mysql version to 8.0.33
* Bump both dockerfiles to 8.0.33 since this version supports Apple
Silicon

* Change JDBC url to support new version

The new version required me to add "allowPublicKeyRetrieval=true"
to jdbc urls in order for it to work.

* Remove is_local and usages of it

The current method for checking if the tests are running locally
did not work on my Apple silicon machine. Per Simon's advice
I removed the check since the tests should always be run in
docker anyway.